+++ Api/LancacheManager/Core/Interfaces/IClientGroupsService.cs
@@ -6,16 +6,16 @@ namespace LancacheManager.Core.Interfaces;
 /// Service interface for client group management.
 /// Extends ICrudRepository for compatibility with CrudControllerBase.
 /// </summary>
-public interface IClientGroupsService : ICrudRepository<ClientGroup, int>
+public interface IClientGroupsService : ICrudRepository<ClientGroup, long>
 {
     // Entity-specific methods
     Task<List<ClientGroup>> GetAllGroupsAsync(CancellationToken cancellationToken = default);
-    Task<ClientGroup?> GetGroupByIdAsync(int id, CancellationToken cancellationToken = default);
+    Task<ClientGroup?> GetGroupByIdAsync(long id, CancellationToken cancellationToken = default);
     Task<ClientGroup?> GetGroupByNicknameAsync(string nickname, CancellationToken cancellationToken = default);
     Task<ClientGroup> CreateGroupAsync(ClientGroup group, CancellationToken cancellationToken = default);
     Task<ClientGroup> UpdateGroupAsync(ClientGroup group, CancellationToken cancellationToken = default);
-    Task DeleteGroupAsync(int id, CancellationToken cancellationToken = default);
-    Task<ClientGroupMember> AddMemberAsync(int groupId, string clientIp, CancellationToken cancellationToken = default);
-    Task RemoveMemberAsync(int groupId, string clientIp, CancellationToken cancellationToken = default);
-    Task<Dictionary<string, (int GroupId, string Nickname)>> GetIpToGroupMappingAsync(CancellationToken cancellationToken = default);
+    Task DeleteGroupAsync(long id, CancellationToken cancellationToken = default);
+    Task<ClientGroupMember> AddMemberAsync(long groupId, string clientIp, CancellationToken cancellationToken = default);
+    Task RemoveMemberAsync(long groupId, string clientIp, CancellationToken cancellationToken = default);
+    Task<Dictionary<string, (long GroupId, string Nickname)>> GetIpToGroupMappingAsync(CancellationToken cancellationToken = default);
 }
